feat(app): registration page redesign
Redesign registration page, rename changed files to Pascal case
