+++ frontend/src/components/common/button/Button.tsx
@@ -1,0 +1,16 @@
+import { ButtonHTMLAttributes, forwardRef } from "react";
+
+export const Button = forwardRef<
+  HTMLButtonElement,
+  ButtonHTMLAttributes<HTMLButtonElement>
+>(({ ...props }, ref) => {
+  return (
+    <button
+      ref={ref}
+      {...props}
+      className={`rounded-md bg-purple-800 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-purple-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-purple-600 ${props.className}`}
+    >
+      {props.children}
+    </button>
+  );
+});
