+++ frontend/src/components/common/button/index.tsx
@@ -1,1 +1,1 @@
-export { default as Button } from "./button";
+export * from "./Button";
