--- frontend/src/pages/register.tsx
@@ -1,5 +1,0 @@
-import { RegisterForm } from "@features/auth";
-
-export const RegisterPage = () => {
-  return <RegisterForm />;
-};
